+++ example/a.ts
@@ -1,12 +1,23 @@
-import { Elysia, t } from '../src'
+import { Elysia, file, form, t } from '../src'
 
-const plugin = new Elysia()
-	.onAfterHandle({ as: 'scoped' }, (x) => {
-		console.log(x.response)
-		return 'From AfterHandle'
-	})
-
-new Elysia()
-	.use(plugin)
-	.get('/', () => 'From Index')
-	.listen(4848)
+/*
+Type 'BunFile[]' is not assignable to type 'File[]'.
+Type 'BunFile' is not assignable to type 'File'.
+Types of property 'name' are incompatible.
+Type 'string | undefined' is not assignable to type 'string'.
+*/
+new Elysia().get(
+	'/test',
+	() => {
+		return form({
+			files: [file('test.png'), file('test.png')],
+			text: 'hello'
+		})
+	},
+	{
+		response: t.Form({
+			files: t.Files(),
+			text: t.String()
+		})
+	}
+)